Re-introduce `two_body` module and more tests
